webkitgtk: support cross compilation
which previously errored near the end of compilation with:
```
/nix/store/nnnn-wayland-aarch64-unknown-linux-gnu-1.22.0-bin/bin/wayland-scanner: line 0: syntax error: unexpected word (expecting ")")
```

that is, the build was trying to invoke the host `wayland-scanner`: both
host and build wayland-scanner are made available to the build
environment and the build script isn't smart enough to choose the right
one. fix by explicitly instructing which `wayland-scanner` to use.