lib.types.anything: Add functionArgs
I'm not a big fan of function reflection, or types.anything, because
both are broken concepts, but this at least fixes the use case where a
function is expected and passed to, say, callPackage or evalModules, etc.

I have chosen not to apply this change to functionTo, because this
solution has a performance cost, and perhaps usages of functionTo are
better - by not using reflection.