parallel-full: fix missing files and meta
Previously, this was not usable as a drop-in replacement for `parallel`
since it only included the one exe. This changes to additionally symlink
all files from $out as well as symlink all other outputs. So no files
are missing compared to `parallel`, while being efficiently symlinked.

Further, copy meta and version from the base package.