mixRelease: remove current attempt for removing erlang references in resulting derivation
As said in the comment about "remove erlang references in resulting
derivation", for now, we don't have a robust method to do that.

Although these removed code did some work, they did not achieve the
final goal - remove erlang references in resulting derivation.
Therefore, it is better to remove them and provide future implementation
with a clean foundation.

> If you want to find these old codes, you can also retrieve them from
> the git history.