darwin.mDNSResponder: patch header to match SDK
The upstream source release for 765.50.9 defines `_DNS_SD_LIBDISPATCH`
to `0` in `dns_sd.h`, but the SDK version of that header actually
defines it to `1`. Having it defined to `0` causes packages that expect
to use libdispatch-based symbols such as `DNSServiceSetDispatchQueue` to
fail to build. Define it to match the SDK to fix those packages.