linux_testing: 6.6-rc7 -> 6.7-rc1
ZSMALLOC is now the default allocator for ZSWAP, so if the allocator
isn't changed, ZSMALLOC must be y on 6.7 (and will default to that).
The script will set all options that can be modules to m anyway, so
there was no need for us to be explicitly setting it to module before
anyway.

(cherry picked from commit 3a7daebe139b6e82f238dc7782b478b3a693ac15)