firebird-emu: support darwin
firebird-emu supports darwin, but the derivation has not been tested on macOS.

This commit adds darwin support to the derivation, with the necessary tweaks to support building for a darwin target.
Namely, adds a workaround for https://github.com/NixOS/nixpkgs/issues/19098, and adds a missing dependency, qtquickcontrols.

Also, adds a postInstall hook which moves the MacOS .app file to the Applications/ folder in the
store, enabling proper linking to /Applications or ~/Applications and appearance in Spotlight
and Launchpad.