emacs: Add withCompressInstall flag
When enabled (the default), Emacs compresses its files (.c, .el, .info,
and so on) before installing them.

Disabling the option with `withCompressInstall = false` results in the
suppression of that compression step. This increases disk space, but is
slightly faster for some operations; Emacs does not have to decompress
these files on the fly when, for example, jumping to the definition of a
built-in function.