papirus-icon-theme: add withElementary optional argument
Upstream advices to make optional the installation of ePapirus and
ePapirus-Dark icon themes, in order to reduce inodes and package
size. Those themes are recommended for the ElementaryOS and Pantheon desktop
environments.

Therefore papirus-icon-theme does not include ePapirus* anymore. But
the new epapirus-icon-theme includes all of them, and is suitable for
use with ElementaryOS or Pantheon.