newlib: evade more of newlib's self-disablement
In https://github.com/NixOS/nixpkgs/pull/266364/files I neglected to
evade one of newlib's other (annoying) self-disablement checks: when
host==target newlib decides it's not in the mood to build itself for
some reason.

Since we have an entire package for newlib (it's no longer part of
gcc) these checks are silly.  So I have evadede the second check as well.

Closes #267859