rustc-wasm32: fix targetPlatform
The previous version stopped working when we started elaborating Rust
metadata.  Here, I've made it a bit nicer by actually setting
targetPlatform to an elaborated system.  Setting the config to wasi to
get elaborate to understand it is a bit of a hack, but I think it's
less of a hack than what we had before.

The only actual difference this makes to the rustc-wasm32 derivation
compared to the previous working version, is that now crt-static is
set.  This is probably the right thing anyway.

Fixes: e3e57b8f1885 ("lib.systems: elaborate Rust metadata")
(cherry picked from commit a578479a42a5d2033cc9fa77f236545f9bec5873)