dvdauthor: 0.7.1 -> 0.7.2
dvdauthor optionally depends on imagemagick6, which is marked insecure.
imagemagick7 could be supported with patches, but since upstream
disabled the feature by default, it seems easier to remove it.
(see https://github.com/ldo/dvdauthor/issues/24)

In addition, the previous release's Nix package listed imagemagick(7) as
dependence, but it was not recognised and not used due to an error in
the dvdauthor build script.