pgbadger: fix build on darwin
The pgbadger build was failing on darwin due to an error while trying to
generate the package's documentation.

The pgbadger script is executed during the build to dump the help text.
The script's shebang line is too long on Darwin and needs to be
shortened using 'shortenPerlShebang' for it to execute.