gitlab: downgrade Ruby from 3.2 to 3.1
Ruby 3.0 was dropped from nixpkgs in #264361 as it will soon become EOL
during the NixOS 23.11 release. As a result, the Ruby version in gitlab
was (accidentally) updated to 3.2.

However, gitlab does not officially support Ruby 3.2 yet.
[gitlab-org&9684] Official support for Ruby 3.1 will be introduced in
the December 2023 release of gitlab. [gitlab-org&10034] Packaging gitlab
with Ruby 3.1 in nixpkgs had been confimed working by a maintainer. As
such, it should be safe to downgrade the Ruby version to 3.1.

[gitlab-org&9684]: https://gitlab.com/groups/gitlab-org/-/epics/9684
[gitlab-org&10034]: https://gitlab.com/groups/gitlab-org/-/epics/10034