maude: remove clang version special casing
Previously, maude would force clang 7 to be used when compiling with a
stdenv that uses some version of clang. The only such case during native
compilation is on Darwin where maude is marked as broken currently.

Thus, I'm treating this as dead code and removing the logic, as it
hinders our ability to remove llvmPackages_7. This may regress some kind
of cross compilation use case, but this can be addressed at some other
point if anyone is relying on that at all.