maintainers/scripts/ec2: drop the AMI creation script
The release management process involves uploading functional AMIs
at each release for NixOS.

As this script mentions, the 'entire thing' should be replaced with a new setup
involving Terraform.

Such a new setup can be found in https://github.com/NixOS/nixos-org-configurations/pull/254.

In the meantime, this script should not be used anymore until a better solution for AMI uploading is
found.

This means effectively that AMI uploading is not part of the release management process for further
releases.

Feel free to push forward the PR mentioned to get back AMI upload.