http-parser: Build on windows
It actually does without modification, but what it installs isn't quite
right. This fixes that.

Specifying `uname` is in general more robust for cross, and specifying
those other variables always (when we just need to for Windows) helps
keep cross/native more similar for maintainability's sake.

The package is no longer maintained, so there isn't really anything to
upstream here.