zfs: backport patches for block cloning tunable and default disable
Due to https://github.com/openzfs/zfs/issues/15533, we do not want to
update to 2.2.1, but we *do* want patches from 2.2.1 to address block
cloning issues (see https://github.com/openzfs/zfs/pull/15529).

The first patch does not apply cleanly due to a trivial conflict in the
context, so vendor the patch.