xdg-desktop-portal: fix disabling Geo location
The "geoclue" Meson option is defined to have type "feature", which
means that the set of possible options is "enabled", "disabled" or
"auto". We pass "false" though when the user has disabled Geo location
support, which is wrong and causes the build to fail:

> meson.build:1:0: ERROR: Value "false" (of type "string") for combo option "Enable Geoclue support. Needed for location portal" is not one of the choices. Possible choices are (as string): "enabled", "disabled", "auto".

This broke in c585543ccd50 (xdg-desktop-portal: 1.16.0 → 1.17.0,
2023-08-08) where we started using the new Meson-based build system
instead of the autotools-based one. Fix it by using "disabled" instead.