qcustomplot: init at 2.1.1
This is a plotting library used by various Qt applications, and is
available in debian. Debian's implementation includes a patch file
that adds cmake files to qcustomplot, allowing it to be built easily.
I have adopted that patch in this commit, and referenced the Debian
source as appropriate.

My own motivation for requiring it in nixpkgs is that I require it
for wfview in a following commit.