cc-wrapper: Enter C++ mode when -lc++ is passed
When linking C++ programs in LLVM we're currently explicitly passing
in -lc++abi as this appears to be required by recent versions of
Clang. However, it also appears to b common for packages to use
the C compiler to link binaries which uses C++, in which case they
do pass in -lc++. This changes the wrapper to also take this into
consideration and enter "C++ mode" whenever -lc++ is passed as an
argument.

Fixes nixos/nixpkgs#166205