i3status-rust: optional icu_calendar support
Recently, upstream has incorporated support for non-Gregorian calendars
using the icu_calendar crate. To manage the substantial increase in
binary size caused by this addition, upstream has made this feature
flag optional. In line with this decision, this commit introduces the
withICUCalendar option, set to a default value of false.