python311Packages.cvxopt: use LP64 blas on darwin; fix build
darwin builds were using the ILP64 openblas library which causes the
build to segufault in the unittests, which might be why the unit tests
were disabled on darwin...

this change aligns darwin to the linux build to use LP64 blas / lapack
libraries. this also aligns with the assert at the top of the file which
asserts blas and lapack are LP64.

the darwin unit test is re-enabled and passes.