nix-doc: broken
`nix-doc` contains a unit test that breaks on current Nixpkgs builds.
As I'm not sure what the test is supposed to check for exactly
and whether the breakage might have been caused by a recent update
of `nix` (the project depends on `nix-expr`), I propose simply marking
it as such for now.

I also opened a corresponding GitHub issue upstream:
https://github.com/lf-/nix-doc/issues/24