paraview: fix strlcat symbol provided by glibc 2.38
Glibc introduces strlcat in commit [1].
The logic of vtknetcdf was to only define strlcat conditionally on not
BSD systems, but starting with glibc 2.38, strlcat is also defined on
Linux.

The CMakelist checks for the symbol existence, the redefinition of the
symbol can be conditional on the boolean HAVE_STRLCAT instead of if
not BSD systems.

This commit fixes 268961.

[1] https://sourceware.org/git/?p=glibc.git;a=commit;h=454a20c8756c9c1d55419153255fc7692b3d2199