teeworlds: fix build on Darwin
The Cocoa dependency is also needed when not building the Client.
The Carbon dependency does not seem to be needed at all.
The libGLU dependency does not currently build on macOS because of one of its transitive dependencies but teeworlds compiles and runs fine without libGLU on macOS, so just make this dependency Linux-only.