libserdes: link with libc++ libc++abi; fix build
when stdenv.cc.libcxx is non null update the libserdes++ libs from
libstd++ to libc++ libc++abi. This fixes the link error on darwin.
However, on x64 linux using stdenv = llvmPackages_16.libcxxStdenv
on linux the build fails unless -std=c++11 or above is used.

testing; built. the only program that depends on libserdes, kcat, links
to the libserdes and not libserdes++.