llvm: Stop using the native llvm-tblgen when buildPlatform can execute hostPlatform
Previously, we change the llvm-tblgen to the native llvm-tblgen and mark
the build as cross compilation if the host and build platform weren't the same.
But some cases that the build platform can execute the host platform binaries,
it will cause a relink of binaries that can lead to a failure. For example
when we're targeting Musl from a glibc build platform.